457: move datastructures to interpret/interpret.js
