+++ src/interpret/setup.js
@@ -7,6 +7,16 @@ import { Point } from '../global/class/Point';
 import { Triangle } from './class/Triangle';
 import { TriangleGraph } from './class/TriangleGraph';
 import { unmergedGraphFromTagproMap, graphFromTagproMap } from './mapToGraph';
+import {
+  getDTGraph,
+  getMergedGraph,
+  getUnmergedGraph,
+  setDtGraph,
+  setUnmergedGraph,
+  setMergedGraph,
+  setMapName,
+  setMapAuthor,
+} from './interpret';
 
 
 export const internalMap = [];
@@ -15,16 +25,11 @@ export const internalMap = [];
 export const tilesToUpdate = [];
 export const tilesToUpdateValues = []; // the values stored in those locations
 
-let dtGraph;
-let unmergedGraph;
-let mergedGraph;
-let mapName;
-let mapAuthor;
 
 export function setupMapCallback() {
   tagpro.socket.on('map', mapData => {
-    mapName = mapData.info.name;
-    mapAuthor = mapData.info.author;
+    setMapName(mapData.info.name);
+    setMapAuthor(mapData.info.author);
   });
 }
 
@@ -69,6 +74,7 @@ export function delaunayTriangulation(
   dummyPoint2,
   dummyPoint3,
 ) {
+  const dtGraph = getDTGraph();
   const numVertices = dtGraph.getVertices().length;
   assert(numVertices === 0, `dtGraph had ${numVertices} vertices.`);
   const vertices = mapGraph.getVertices();
@@ -102,30 +108,15 @@ export function delaunayTriangulation(
  * @returns {Graph} graph of the triangulation of all the vertices
  */
 export function initNavMesh(map) {
-  dtGraph = new TriangleGraph();
+  setDtGraph(new TriangleGraph());
   timeLog('  Creating unmerged graph...');
-  unmergedGraph = unmergedGraphFromTagproMap(map);
+  setUnmergedGraph(unmergedGraphFromTagproMap(map));
   timeLog('  Creating merged graph...');
-  mergedGraph = graphFromTagproMap(map, unmergedGraph);
+  setMergedGraph(graphFromTagproMap(map, getUnmergedGraph()));
   delaunayTriangulation(
-    mergedGraph,
+    getMergedGraph(),
     new Point(-9999, -100),
     new Point(9999, -100),
     new Point(0, 9999),
   );
-}
-
-
-export function getMergedGraph() {
-  return mergedGraph;
-}
-
-
-export function getUnmergedGraph() {
-  return unmergedGraph;
-}
-
-
-export function getDTGraph() {
-  return dtGraph;
 }
